pyproject.toml
[build-system]
requires = ["setuptools>=72.0"]
build-backend = "setuptools.build_meta"

[project]
name = "spritekit"
version = "0.1.2"
authors = [
  { name="George Watson", email="gigolo@hotmail.co.uk" },
]
description = "2D scene+actor framework built on top of raylib"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "transitions==0.9.2",
  "pyglm==2.8.1",
  "raudio==0.0.1",
  "quickwindow==0.1.0",
  "moderngl==5.12.0",
  "numpy==2.2.4",
  "pillow==11.1.0",
  "pyglsl==0.1.0"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/takeiteasy/PySpriteKit"
Issues = "https://github.com/takeiteasy/PySpriteKit/issues"
